+++ includes/minishell.h
@@ -125,7 +125,6 @@ void ft_unset(t_node *node);
 char **ft_split(char *s, char c);
 void print_all_envp(t_env *envp, int has_prefix);
 void child_sig_int(int signal);
-void heredoc_sig_int(int signal);
 void print_sorted_envp(t_env *envp);
 t_env *update_env(t_env *envp, char *key, char *value);
 int is_duplicate_envp(t_env *envp, char *key);
@@ -142,6 +141,13 @@ char *find_cmd_path(char *cmd, char **paths);
 t_list *tmp_files(char *filename, int cmd);
 t_list *add_files(t_list *head, char *filename);
 void delete_files(t_list *head);
+int get_heredoc_fd(t_node *node);
+void heredoc_sig_int(int signal);
+
+// fd
+int set_input_fd(t_node *head, t_node *file_node);
+int set_output_fd(t_node *head, t_node *file_node);
+int set_separator_fd(t_node *node);
 
 // parse error
 t_node *error_handler(t_node *head, t_token **token, t_token **tmp);
