+++ main.c
@@ -19,21 +19,20 @@ int g_stat;
 void sig_int(int signal)
 {
 	if (signal != SIGINT)
-		return;
+		return ;
 	g_stat = ETC;
 	ioctl(STDIN_FILENO, TIOCSTI, "\n");
-	rl_on_new_line();		// 개행이 나온 후 수행되어야함
-	rl_replace_line("", 1); // 버퍼 비우고
-							// rl_redisplay();         // prompt
+	rl_on_new_line();
+	rl_replace_line("", 1);
 }
 
 void heredoc_sig_int(int signal)
 {
 	if (signal != SIGINT)
-		return;
+		return ;
 	ioctl(STDIN_FILENO, TIOCSTI, "\n");
-	rl_on_new_line();		// 개행이 나온 후 수행되어야함
-	rl_replace_line("", 1); // 버퍼 비우고
+	rl_on_new_line();
+	rl_replace_line("", 1);
 	g_stat = ETC;
 }
 
@@ -69,105 +68,6 @@ t_env *init_env(char **env)
 		i++;
 	}
 	return (lst);
-}
-
-// test code
-void print_token(t_token *token, int flag)
-{
-	char type[7][10] = {"CMD", "PIPE", "TRUNC", "APPEND", "INPUT", "HEREDOC", "END"};
-	while (token && flag == 1)
-	{
-		printf("[%s : %s]", token->value, type[token->type - 1]);
-		token = token->nxt;
-	}
-	while (token && flag == 0)
-	{
-		printf("[%s]", token->value);
-		token = token->nxt;
-	}
-	printf("\n");
-}
-
-// test code
-void print_node(t_node *node)
-{
-	int i;
-	char type[7][10] = {"CMD", "PIPE", "TRUNC", "APPEND", "INPUT", "HEREDOC", "END"};
-
-	while (node)
-	{
-		i = 0;
-		printf("-------- %s ----------\n", type[node->type - 1]);
-		while (node->cmd && node->cmd[i])
-		{
-			printf("%s\n", node->cmd[i]);
-			i++;
-		}
-		printf("infile : %d\n", node->fd[IN]);
-		printf("outfile : %d\n", node->fd[OUT]);
-		node = node->nxt;
-	}
-}
-
-// test code
-void print_heredoc(t_node *node)
-{
-	char *line;
-
-	if (!node)
-		return;
-
-	while (node)
-	{
-		if (node->type == HEREDOC)
-		{
-			line = get_next_line(node->fd[IN]);
-			printf("%d\n", node->fd[IN]);
-			printf("------- heredoc -------\n");
-			while (line)
-			{
-				printf("line : %s\n", line);
-				free(line);
-				line = get_next_line(node->fd[IN]);
-			}
-		}
-		if (node)
-			node = node->nxt;
-	}
-}
-
-// test code
-void print_tmpfiles()
-{
-	t_list *tmp = tmp_files(NULL, GET);
-	printf("------ tmps files ------\n");
-	while (tmp)
-	{
-		printf("%s\n", tmp->value);
-		tmp = tmp->nxt;
-	}
-}
-
-void check_redirection(t_node *node)
-{
-	t_node *tmp;
-
-	while (node)
-	{
-		if (node->type == TRUNC || node->type == APPEND)
-		{
-			tmp = node->nxt;
-			while (tmp)
-			{
-				if (tmp->type == PIPE)
-					break;
-				if (tmp->type == CMD)
-					tmp->fd[OUT] = node->fd[OUT];
-				tmp = tmp->nxt;
-			}
-		}
-		node = node->nxt;
-	}
 }
 
 int main(int ac, char **av, char **env)
@@ -239,9 +139,6 @@ int main(int ac, char **av, char **env)
 			free_node_all(node);
 			tmp_files(NULL, DEL);
 		}
-		// if (tmp_files(NULL, GET) == NULL)
-		//     printf("yes\n");
-		// printf("stat : %d\n", g_stat);
 	}
 	return (0);
 }
