+++ srcs/parse_utils/parse_error.c
@@ -1,4 +1,4 @@
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 extern int g_stat;
