+++ srcs/signals/signal.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 #include <sys/ioctl.h>
 #include <termios.h>
