+++ srcs/token/split_by_sep.c
@@ -1,4 +1,4 @@
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 extern int g_stat;
